refactor to modal component
