+++ nr-app/app/(tabs)/settings.tsx
@@ -7,15 +7,9 @@ import { setVisiblePlusCodes } from "@/redux/actions/map.actions";
 import { useAppDispatch, useAppSelector } from "@/redux/hooks";
 import { setPrivateKeyMnemonicPromiseAction } from "@/redux/sagas/keystore.saga";
 import { keystoreSelectors } from "@/redux/slices/keystore.slice";
-import {
-  Button,
-  Modal,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-} from "react-native";
-import { getNip5PubKey } from "../../../nr-common/utils.ts";
+import { Button, Modal, StyleSheet, Text, View } from "react-native";
+
+import OnboardModal from "@/components/OnboardModal";
 
 import {
   settingsActions,
@@ -23,16 +17,7 @@ import {
 } from "@/redux/slices/settings.slice";
 import { generateSeedWords, getBech32PrivateKey } from "nip06";
 import { useEffect, useState } from "react";
-import {
-  Linking,
-  SafeAreaView,
-  ScrollView,
-  Switch,
-  TextInput,
-} from "react-native";
-import Toast from "react-native-root-toast";
-
-import * as Clipboard from "expo-clipboard";
+import { SafeAreaView, ScrollView, Switch, TextInput } from "react-native";
 
 const DevSwitch = () => {
   const dispatch = useAppDispatch();
@@ -61,12 +46,6 @@ const DevSwitch = () => {
 export default function TabThreeScreen() {
   const username = useAppSelector(settingsSelectors.selectUsername);
 
-  const [text, setText] = useState<string>("");
-  const [error, setError] = useState<string | null>(null);
-  const [nip5VerificationLoading, setNip5VerificationLoading] =
-    useState<boolean>(false);
-
-  const [modalVisible, setModalVisible] = useState(false);
   const [nsec, setNsec] = useState("");
   const [mnemonic, setMnemonic] = useState("");
   const hasPrivateKeyFromRedux = useAppSelector(
@@ -87,132 +66,30 @@ export default function TabThreeScreen() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const OpenTrustRootsButton = () => {
-    const url = "https://www.trustroots.org/";
-    const handlePress = async () => {
-      const supported = await Linking.canOpenURL(url);
-      if (supported) {
-        await Linking.openURL(url);
-      } else {
-        // fail silently, some kind of compatibility error happened?
-        console.error(`Don't know how to open this URL: ${url}`);
-      }
-    };
-
-    return <Button title="Open TrustRoots Website" onPress={handlePress} />;
-  };
-
-  const handleUsernameSubmit = async () => {
-
-    // sanity check if the username exists
-    if (username !== "") {
-      return;
-    }
-
-    if (text.trim().length < 3 || !/^[a-zA-Z0-9]+$/.test(text.trim())) {
-      setError("Input must be at least 3 alphanumeric characters.");
-      return;
-    }
-    // freeze the text box
-    setNip5VerificationLoading(true);
-    const nip5Result = await getNip5PubKey(text);
-
-    setNip5VerificationLoading(false);
-    if (nip5Result === undefined) {
-      setError("Invalid username for this key.");
-      return;
-    }
-    setError(null);
-
-    setText("");
-    dispatch(settingsActions.setUsername(text));
-  };
-
-  const handleTextChange = (newText: string) => {
-    if (nip5VerificationLoading === false) {
-      setText(newText);
-    }
-  };
-
-  const handleCopy = async (npub) => {
-    await Clipboard.setStringAsync(npub || "");
-    Toast.show("Copied public key to Clipboard!", {
-      duration: Toast.durations.LONG,
-      position: Toast.positions.TOP,
-    });
-  };
-
-  const modal = (
-    <Modal
-      animationType="slide"
-      transparent={false}
-      visible={modalVisible}
-      onRequestClose={() => setModalVisible(false)}
-    >
-      <View style={styles.modalContainer}>
-        <View style={styles.instructions}>
-          <Text style={styles.modalText}>
-            How to link your app key to trustroots.org
-          </Text>
-          <Text style={styles.modalText}>1) Log into trustroots.org</Text>
-          <Text style={styles.modalText}>
-            2) Copy and save your public key on trustroots
-          </Text>
-          <Text style={styles.modalText}>
-            3) Save your trustroots user name in the app.
-          </Text>
-
-          <View style={styles.usernameContainer}>
-            <Text style={styles.usernameLabel}>Enter text:</Text>
-            <TextInput
-              style={[
-                styles.usernameInput,
-                error ? styles.usernameInputError : null,
-              ]}
-              value={text}
-              onChangeText={handleTextChange}
-              placeholder="Verify your trustroots username"
-            />
-            {error && <Text style={styles.error}>{error}</Text>}
-            <Text style={styles.output}>You typed: {text}</Text>
-
-            <TouchableOpacity
-              style={styles.usernameSubmitBtn}
-              onPress={handleUsernameSubmit}
-            >
-              <Text style={styles.usernameSubmit}>Submit</Text>
-            </TouchableOpacity>
-          </View>
-
-          <OpenTrustRootsButton />
-          <Button
-            title="Copy Public Key to Clipboard"
-            onPress={() => handleCopy(npub)}
-          />
-        </View>
-
-        <TouchableOpacity
-          style={styles.closeButton}
-          onPress={() => setModalVisible(false)}
-        >
-          <Text style={styles.closeButtonText}>Close</Text>
-        </TouchableOpacity>
-      </View>
-    </Modal>
-  );
+  const [modalVisible, setModalVisible] = useState(false);
 
   return (
     <SafeAreaView style={styles.settings}>
       <ScrollView>
-        <Text style={styles.header}>username: {username}</Text>
-        <Text style={styles.header}>Keys</Text>
         <View>
+          <Modal
+            animationType="slide"
+            transparent={false}
+            visible={modalVisible}
+            onRequestClose={() => setModalVisible(false)}
+          >
+            <OnboardModal npub={npub} setModalVisible={setModalVisible} />
+          </Modal>
+        </View>
+        {username === "" ? (
           <Button
             title="Link your TrustRoots Key"
             onPress={() => setModalVisible(true)}
           />
-          {modal}
-        </View>
+        ) : (
+          <Text style={styles.q}>username: {username}</Text>
+        )}
+
         <Text style={styles.q}>npub</Text>
         <TextInput style={styles.input} value={npub} />
         {/**
